requirements.txt
anthropic==0.34.2
faiss-cpu==1.8.0
Flask==3.0.3
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.0
openai==1.48.0
python-dotenv==1.0.1
pandas